requirements.txt

python-dotenv
notion-client
flask
flask-dotenv 